Remove dependency on argument names in the value source generator
It is valid C# to have different names to the expected ones on the interface.
If we depend on the names we end up generating conflicting members if the
argument name is not what we were expecting, leading to compiler errors.

This also increases the precision of the signature checks to make sure we can
disambiguate `char` and `byte` overloads by checking the generic argument
type.